Resolve merge conflicts, keep hem4
