+++ README.md
@@ -1,2 +1,5 @@
-# hem2
-hem 2
+# hem4
+hem 4
+
+This repository contains a single script `hem4.py`, verified to have no merge conflicts.
+
